added 2d to 3d processing and filtering

+++ app/ThruMu/BMTrackCluster3D.h
@@ -19,14 +19,15 @@ namespace larlitecv {
     float tick_end;
     std::vector<int> start_wire;
     std::vector<int> end_wire;
-    std::vector<float> start3D;
-    std::vector<float> end3D;
+    std::vector<double> start3D;
+    std::vector<double> end3D;
     std::vector< BoundaryEndPt > start_endpts;
     std::vector< BoundaryEndPt > end_endpts;
     BoundaryEndPt::BoundaryEnd_t start_type;
     BoundaryEndPt::BoundaryEnd_t end_type;
     std::vector< BMTrackCluster2D > plane_paths;
-    std::vector< std::vector<float> > path3d;
+    std::vector< std::vector<double> > path3d;
+    
     //std::vector< float > node_triangle_area;
 
     bool operator< (const BMTrackCluster3D& rhs ) const {
